removes form from index.erb
this form should be added after running the second test and having it fail.  As it is, the test passes immediately once written because the form is already inside of index.erb.  Instructions for adding the form to index.erb are in the README.md, so there is no need for the code to already be inside of index.erb.