removes HTML for greeting in greet.erb
The Third test passes automatically when it is written because this code is already in the greet.erb file.  The instructions for adding this code to this file are in the README.md, so there is no need for them to already be here.  We will be able to follow the instructions as written if this code is not in the greet.erb file in the repository upon cloning.